feat(levels.ts): add module for log levels and related utilities
The `levels.ts` module introduces an enumeration `LogLevel` that represents different log levels. It also defines a type `LogLevelName` to represent the names of log levels. The module includes an array `levelNames` that contains the names of log levels derived from the `LogLevel` enumeration.

Additionally, the module provides a mapping `levelByName` that maps level values to their corresponding names. This mapping is used to retrieve the name of a log level given its value.

The module includes two utility functions: `getLevelByName` and `getLevelName`. `getLevelByName` takes a `LogLevelName` as input and returns the corresponding `LogLevel` value. It throws an error if the provided level name is unknown. `getLevelName` takes a log level value as input and returns the name of the log level. It also throws an error if the provided level is unknown.

Lastly, the module defines a function `colorizeLevel` that colorizes the level of a log message. It takes a `LogLevel` value and a color as input and returns the colorized log message level.

These additions improve the organization and reusability of log level-related functionality in the codebase.