feat(colors.ts): add colors module to enable formatting in the application
The colors module is added to the project to enable formatting. It creates and returns a colors object that can be used for formatting text in the application. The `createColors` function from the `picocolors` library is used to create the colors object. The `enableFormatting` parameter is set to `true` to enable formatting.