feat(levels.ts): add coloredLevel object and getColoredLevel function to provide colored representation of log levels
The coloredLevel object is a mapping of log levels to their corresponding colors. This allows for easy customization of the colors used for each log level.

The getColoredLevel function takes a log level as input and returns the colored representation of that log level. This can be used to display log messages with the appropriate color based on their level.